+++ KiLib/Raster/Raster.hpp
@@ -24,8 +24,10 @@
 #include <algorithm>
 #include <cmath>
 #include <fstream>
+#include <functional>
 #include <iomanip>
 #include <iostream>
+#include <map>
 #include <random>
 #include <sstream>
 #include <string>
@@ -65,6 +67,22 @@ namespace KiLib
       {
          KiLib::Raster new_(other);
          std::fill(new_.data.begin(), new_.data.end(), 0.0);
+         return new_;
+      }
+
+      // Creates a raster with same metadata as other, filled with fillValue.
+      // If keepNoData is true, returned raster will have nodata in same locations as other.
+      // Otherwise every value will be fillValue
+      static KiLib::Raster fillLike(const KiLib::Raster &other, double fillValue, bool keepNoData)
+      {
+         KiLib::Raster new_(other);
+         for (size_t i = 0; i < other.nData; i++)
+         {
+            if (new_(i) != new_.nodata_value)
+            {
+               new_(i) = fillValue;
+            }
+         }
          return new_;
       }
 
@@ -226,6 +244,56 @@ namespace KiLib
 
       KiLib::Raster        ComputeSlope(KiLib::Raster::SlopeMethod method) const;
       static KiLib::Raster ComputeSlopeZevenbergenThorne(const KiLib::Raster &inp);
+
+      /**
+       * @brief Takes in a vector of objects, and takes the mean of a given attribute at each cell position in a raster.
+       * The attributes and corresponding positions are mapped to the nearest cell in the raster, and the mean is taken
+       * over cells.
+       *
+       * @tparam T obj
+       * @param ref Reference raster to determine shape, size, nodata, etc
+       * @param objs vector of objects
+       * @param posP Pointer to position attribute (i.e. &Landslide::pos)
+       * @param attrP Pointer to attribute to rasterize (i.e. &Landslide::safetyFactor)
+       * @param doPopulate function that dictates whether or not that instace will be rasterized
+       * @return KiLib::Raster
+       *
+       */
+      template <class T>
+      static KiLib::Raster Rasterize(
+         const KiLib::Raster &ref, const std::vector<T> &objs, KiLib::Vec3 T::*posP, double T::*attrP,
+         std::function<bool(const T &)> doPopulate)
+      {
+         KiLib::Raster            outRast = KiLib::Raster::fillLike(ref, 0.0, true);
+         std::map<size_t, double> counts;
+
+         // Set up getters for the attributes
+         auto getPos  = std::bind(posP, std::placeholders::_1);
+         auto getAttr = std::bind(attrP, std::placeholders::_1);
+
+         for (auto &obj : objs)
+         {
+            if (!doPopulate(obj))
+            {
+               continue;
+            }
+
+            KiLib::Vec3 pos  = getPos(obj);
+            double      attr = getAttr(obj);
+
+            size_t flatInd  = outRast.getNearestCell(pos);
+            counts[flatInd] = counts.count(flatInd) == 0 ? 1.0 : counts[flatInd] + 1;
+
+            outRast.at(flatInd) += attr;
+         }
+
+         for (auto [ind, count] : counts)
+         {
+            outRast.at(ind) /= count;
+         }
+
+         return outRast;
+      }
 
    private:
       void fromDEM(const std::string path);