+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Notable changes to this project will be documented in this file.
 ### Added
 
 - Add flattenIndex method to Raster
+- Add Rasterize method to raster
 
 ### Fixed